requirements.txt
streamlit>=1.33.0
openai>=1.12.0
python-dotenv>=1.0.1
